+++ rocketgram/api/answer_inline_query.py
@@ -7,6 +7,7 @@ from dataclasses import dataclass
 from typing import Optional, Tuple
 
 from .inline_query_result import InlineQueryResult
+from .inline_query_results_button import InlineQueryResultsButton
 from .request import Request
 from .utils import BoolResultMixin
 
@@ -23,5 +24,4 @@ class AnswerInlineQuery(BoolResultMixin, Request):
     cache_time: Optional[int] = None
     is_personal: Optional[bool] = None
     next_offset: Optional[str] = None
-    switch_pm_text: Optional[str] = None
-    switch_pm_parameter: Optional[str] = None
+    button: Optional[InlineQueryResultsButton] = None