+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project.
 
 ### Added
 - The `commonfilters.gamne` filter.
+- The `InlineQueryResultsButton` class and `button` field for `AnswerInlineQuery` class.
 
 ### Changed.
 - The parameter `executor` is now optional for `init`, `shutdown` and `process` methods of `Bot` class.
@@ -15,6 +16,9 @@ All notable changes to this project.
 ### Fixed
 - Typing in `Animation`.
 - The `commonfilters.callback` filter for case where `game` event has arrived.
+
+### Removed
+- `switch_pm_text` and `switch_pm_parameter` fields from `AnswerInlineQuery` class.
 
 ### Deprecated
 - The next major release will no longer officially support Python 3.7.